create RelayTest.cs
Transfer code from TestRelay.cs to RelayTest.cs, change game object TestRelay to RelayTest
