+++ Assets/Scripts/RelayTest.cs
@@ -1,0 +1,81 @@
+using System.Collections;
+using System.Collections.Generic;
+using QFSW.QC;
+using Unity.Netcode;
+using Unity.Netcode.Transports.UTP;
+using Unity.Networking.Transport.Relay;
+using Unity.Services.Authentication;
+using Unity.Services.Core;
+using Unity.Services.Lobbies;
+using Unity.Services.Lobbies.Models;
+using Unity.Services.Relay;
+using Unity.Services.Relay.Models;
+using UnityEngine;
+using System.Threading.Tasks;
+using Cinemachine;
+
+public class RelayTest : MonoBehaviour
+{
+    public static RelayTest Instance;
+
+    private void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    private void Update()
+    {
+        
+    }
+
+    private void Awake() {
+        if (Instance != null)
+            Destroy(Instance);
+        else
+            Instance = this;
+    }
+
+    [Command]
+    public async Task<string> CreateRelay() {
+        try {
+            Allocation allocation = await RelayService.Instance.CreateAllocationAsync(3);
+
+            string joinCode = await RelayService.Instance.GetJoinCodeAsync(allocation.AllocationId);
+
+            Debug.Log(joinCode);
+
+            RelayServerData relayServerData = new RelayServerData(allocation, "dtls");
+
+            NetworkManager.Singleton.GetComponent<UnityTransport>().SetRelayServerData(relayServerData);
+
+            NetworkManager.Singleton.StartHost();
+
+            GameManager.PlayerJoined(EditPlayerName.Instance.GetPlayerName());
+
+            return joinCode;
+        } catch (RelayServiceException e) {
+            Debug.Log(e);
+            return null;
+        }
+    }
+
+    [Command]
+    public async void JoinRelay(string joinCode) {
+        try {
+            Debug.Log("Joining Relay with " + joinCode);
+            JoinAllocation joinAllocation = await RelayService.Instance.JoinAllocationAsync(joinCode);
+
+            RelayServerData relayServerData = new RelayServerData(joinAllocation, "dtls");
+
+            NetworkManager.Singleton.GetComponent<UnityTransport>().SetRelayServerData(relayServerData);
+
+            NetworkManager.Singleton.StartClient();
+
+            GameManager.PlayerJoined(EditPlayerName.Instance.GetPlayerName());
+
+        } catch (RelayServiceException e) {
+            Debug.Log(e);
+        }
+    }
+}
